requirements.txt
aiogram==3.4.0
aiohttp==3.9.1
schedule==1.2.2
python-dotenv==1.0.0
asyncio==3.4.3